Allow saving the streamer token
Alright, so the issue with listening to onStreamOnline is that it
required the STREAMER'S token, not just any token. The reason I didn't
catch this in testing is because my bot's name matched the channel name,
but in my AdamLearnsLive channel, the bot's name is AdamLearnsBot.

Thankfully, a lot of the code looks cleaner as a result of these
changes. E.g. `init.ts` no longer makes a BotDatabase, and the database
code to fetch tokens now takes an ID as you might expect.