Split single package into a workspace
Ideally, this code is essentially identical in function to what I
already had, but I installed slightly different dependencies, so at the
very least, I know that I can no longer specify "localhost" in the
database connection string and must now use "127.0.0.1".

This is all in preparation for adding a web interface to the bot.