Add dark mode toggle icon and preference saving
Added dark mode toggle button to the top right of the page.

The icon of the button depends on the current theme/mode:
Click on a Sun -> Light Mode, click on a Moon -> Dark Mode.

The user's preference is stored in Local Storage and used each time the
page is loaded to display the appropriate mode