Truncate YouTube messages
YouTube only allows 200 characters, and I don't want to keep creating
commands that can only fit within 200 characters.

Now, when a command is too big, we'll link to it on a.bot.land and then
truncate a preview of it. Quotes will also be truncated.

I also fixed a bug where YouTube's "reply" function didn't actually have
any message information, so it couldn't tag the author.