Add April Fool's game
I made a custom .env.aprilfools.local that would log in with the test
database but under AdamLearnsLive, then I ran `NODE_ENV=aprilfools pnpm
start` in the "bots" package and started the web package. After that, I
just did `!startgame` from my chat to kick everything off.

This code was written hastily in a day, so it's not the best quality.
Saving for posterity.